Download first preference voting data from each polling booth, from the six Australian Federal elections between 2001 and 2016.

Download first preference voting data from each polling booth, from the six Australian Federal elections between 2001 and 2016.

firstpref_pollingbooth_download(...)

Format

A data frame with the following variables:

StateAb: Abbreviation for state name

DivisionID: Electoral division ID

DivisionNm: Electoral division name

PollingPlaceID: Polling place ID

PollingPlace: Polling place name

CandidateID: Candidate ID

Surname: Candidate surname

GivenNm: Candidate given name

BallotPosition: Candidate's position on the ballot

Elected: Whether the candidate was elected (Y/N)

HistoricElected: Whether the candidate is the incumbent member

PartyAb: Abbreviation for political party name

PartyNm: Political party name

OrdinaryVotes: Number of ordinary votes cast at the polling place for the candidate

Swing: Percentage point change in ordinary votes for the party from the previous election

PremisesPostCode: Post code of polling booth

Latitude: Coordinates

Longitude: Coordinates

year: Election year

Arguments

...

Additional arguments passed to `download.file`

Downloads and returns first preference votes for candidates in the House of Representatives, for each polling both, in the seven Australian Federal elections between 2001 and 2016.

Value

A data frame containing first preference votes

A dataset containing first preference vote counts, candidate names, polling place locations, and other results for the House of Representatives from the 2001, 2004, 2007, 2010, 2013 and 2016 Australian federal elections. This data set is obtained using the `firstpref_pollingbooth_download` function. The data were obtained from the Australian Electoral Commission.

Examples

if (FALSE) {
fp_pp <- firstpref_pollingbooth_download()
library(dplyr)
fp_pp %>% filter(year == 2016) %>% arrange(-OrdinaryVotes) %>% head
}