Load shapefile of Australia into R

This is only used to create the new electorate map data, for future years. Load shapefile into R as an sf (simple features) object, extract polygon information, thin polygon, fix any problematic polygons, and format variable names. "nat_map" and "nat_data" objects for every Australian federal election between 2001-2019 can be readily loaded from the package for analysis.

load_shapefile(path_to_shapeFile, tolerance = 0.005)

Arguments

path_to_shapeFile

path to object in local machine

tolerance

numerical tolerance value to be used by the Douglas-Peuker algorithm

Value

object of class SpatialPolygonsDataFrame

Details

The function will take several minutes to complete.

Examples

if (FALSE) {
# Load electorate shapefile into R

# Path to your shapefile
fl <- "local/path/to/shapefile.shp"

# Load
my_sF <- load_shapefile(fl)
}