Data and centroids corresponding to the Australian Electorates from 2016

A dataset containing some demographic information for each of the 150 Australian electorates. The data were obtained from the Australian Electoral Commission, and downloaded from http://www.aec.gov.au/Electorates/gis/gis_datadownload.htm. Must be loaded using the `nat_data_download` function.

nat_data16

Format

A data frame with 150 rows with the following variables:

id: Numeric identifier for the polygon

elect_div: Electorate division name

state: abbreviation of the state name

numccds: AEC variable that might be filled with meaning or a description down the road

area_sqkm: combined square kilometers of each electorate

long_c: longitude coordinate of electorate (polygon) centroid

lat_c: latitude coordinate of electorate (polygon) centroid

x: latitude coordinate for plotting a cartogram

y: longitude coordinate for plotting a cartogram

radius: variable used in the construction of cartogram points