Two candidate preferred votes for candidates (House of Representatives) in each electorate.

A dataset containing two candidate preferred vote counts, and other results for the House of Representatives from the 2001 to 2019 Australian federal elections. Includes the count of votes for the leading two candidates in the electorate after distribution of preferences.

tcp01

tcp04

tcp07

tcp10

tcp13

tcp16

tcp19

tcp22

Format

A data frame with the following variables:

For the 2001 election:

UniqueID

Numeric identifier that links the electoral division with Census and other election datasets

StateAb

Abbreviation for state name

DivisionNm

Electoral division name

Surname

Candidate surname

GivenNm

Candidate given name

Elected

Whether the candidate was elected (Y/N)

Percent

Percentage of ordinary votes cast for the candidate

PartyAb

Abbreviation for political party name

PartyNm

Political party name

Swing

Percentage point change in ordinary votes for the party from the previous election

For the 2004 election:

StateAb

Abbreviation for state name

UniqueID

Numeric identifier that links the electoral division with Census and other election datasets.

DivisionNm

Electoral division name

BallotPosition

Candidate's position on the ballot

CandidateID

Candidate ID

Surname

Candidate surname

GivenNm

Candidate given name

PartyAb

Abbreviation for political party name

PartyNm

Political party name

Elected

Whether the candidate was elected (Y/N)

OrdinaryVotes

Number of ordinary votes cast for the candidate

Percent

Percentage of ordinary votes cast for the candidate

For the 2007 election onwards:

StateAb

Abbreviation for state name

UniqueID

Numeric identifier that links the electoral division with Census and other election datasets.

DivisionNm

Electoral division name

BallotPosition

Candidate's position on the ballot

CandidateID

Candidate ID

Surname

Candidate surname

GivenNm

Candidate given name

PartyAb

Abbreviation for political party name

PartyNm

Political party name

Elected

Whether the candidate was elected (Y/N)

HistoricElected

Whether the candidate is the incumbent member

OrdinaryVotes

Number of ordinary votes cast for the candidate

Percent

Percentage of ordinary votes cast for the candidate

Source

https://www.aec.gov.au/About_AEC/Publications/statistics/files/aec-2001-election-statistics.zip.

https://results.aec.gov.au/12246/results/HouseDownloadsMenu-12246-csv.htm

https://www.aec.gov.au/elections/federal_elections/2004/downloads.htm.

https://results.aec.gov.au/13745/Website/HouseDownloadsMenu-13745-csv.htm

https://www.aec.gov.au/elections/federal_elections/2007/downloads.htm.

https://results.aec.gov.au/15508/Website/HouseDownloadsMenu-15508-csv.htm

https://www.aec.gov.au/elections/federal_elections/2010/downloads.htm

https://results.aec.gov.au/17496/Website/HouseDownloadsMenu-17496-csv.htm

https://www.aec.gov.au/elections/federal_elections/2013/downloads.htm

https://results.aec.gov.au/20499/Website/HouseDownloadsMenu-20499-csv.htm

https://www.aec.gov.au/elections/federal_elections/2016/downloads.htm

https://results.aec.gov.au/24310/Website/HouseDownloadsMenu-24310-Csv.htm

https://results.aec.gov.au/27966/Website/HouseDownloadsMenu-27966-Csv.htm