Download two party preference voting data from each polling booth, from the seven Australian Federal elections between 2001 and 2016.

Download two party preference voting data from each polling booth, from the seven Australian Federal elections between 2001 and 2016.

twoparty_pollingbooth_download(...)

Format

A data frame with the following variables:

StateAb: Abbreviation for state name

DivisionID: Electoral division ID

DivisionNm: Electoral division name

PollingPlaceID: Polling place ID

PollingPlace: Polling place name

LNP_Votes: Count of two party preferred vote in favour of the Liberal National coalition

LNP_Percent: Percentage of two party preferred vote in favour of the Liberal National coalition

ALP_Votes: Count of two party preferred vote in favour of the Labor party

ALP_Percent: Percentage of two party preferred vote in favour of the Labor party

TotalVotes: Total number of votes cast

Swing: Percentage point change in two party preferred vote from the previous election

PremisesPostCode: Post code of polling booth

Latitude: Coordinates

Longitude: Coordinates

year: Election year

Arguments

...

Additional arguments passed to `download.file`

Downloads and returns the two party preferred votes for candidates in the House of Representatives, for each polling both, in the six Australian Federal elections between 2001 and 2016.

Value

A data frame containing two party preference votes

A dataset containing two party preferred vote counts, winning candidate names, polling place locations, and other results for the House of Representatives from each of the 2001, 2004, 2007, 2010, 2013 and 2016 Australian federal elections. Includes the count of votes for the Australian Labor Party and the count of votes for the Liberal-National Coalition for each polling place. This data set is obtained using the `twoparty_pollingbooth_download` function. The data were obtained from the Australian Electoral Commission.

Examples

if (FALSE) {
tpp_pp <- twoparty_pollingbooth_download()
library(dplyr)
tpp_pp %>% filter(year == 2016) %>% arrange(-LNP_Percent) %>% head
}